Be a little more principled in how we provide data to odometry_fusion.
IMU provides reliable roll and pitch, and also 3-axis accelerations.

It should also be reliable for differential yaw (vy), but not absolute.

How this squares with /vo also giving absolute orientation information is unclear.

Perhaps this is not principled enough, though--we still encounter situations where /vo *should* be ignored (white wall, or just tapped or collided more forcefully with a wall somewhere), but it isn't, or something, and the localization starts drifting catastrophically.